Add in class for Bat enemy
Created a class for the bat enemy. There is a lot of methods to be implemented later when we need enemies to have more functionality than just moving around the screen. I also got rid of the enemy state interface, just since I would assume we just use the state machine that has already been created.