Add Blade Trap
I added it as an IEnemy, which may seem counterintuitive, however this item will need to act as an enemy when the player collides with it, ie: taking damage or being destroyed by the player. If anyone has an alternative suggestion, lmk! I would assume this classification doesn't matter too much but I thought to explain myself since it is not clear via the implementation!