Enemy collision overhaul
In this PR:
- Refactored dodongo
   - I created a new dodongo class that is similar to goriya since the old one using the state pattern caused a lot of development pain
   - the issue with the sprites was that instead of including the initialization of the sprite in the spawn method, this needed to be in the constructor. simple fix but it gets rid of the null reference exception :)
   - If you don't like it, please let me know. If you do like it, I can remove the old dodongo code in a proceeding PR
- Changed enemy death effect and spawn effect
   - removed sprites for the enemy cloud, since essentially its the same as an explosion and the sprite sequence we had for explosion looked better anyway
   - Utilizing the new timer utility for simplification of the classes overall. I believe this makes them more intuitive to understand, and less complex by removing an inheritence of iupdateable
   - created a new interface for ienemyeffect which may help Ethan in his development of the level building
   - created an aquamentus ball effect to mimic explosions when the fireballs collide with objects
- Refactored collisions for enemies that use the state machine, and those that do not as well.
   - Please use the enemy cycler to test these collisions. Again, if I forgot any interactions let me know.
   - the only thing that I can think of to fix here with these, is sometimes there is a little bit of jumping with sprites when a collision occurs. how might I keep this from happening?

There is a lot in this PR, which may make it harder to review. I understood this going in, but there is a lot of work that needed to be done here, represented by the issues on the board. If you need clarification on anything let me know, I will be more than happy to explain and improve upon anything here. My plan going forward is to clean up some of the code with the enemies by making utilities that help with the randomization of the movement, since the randomness isnt exactly random, since it uses the same seed every time.