Enemy Flashing / Damage Update
All enemy health has been updated according to the values on https://gamefaqs.gamespot.com/boards/563433-the-legend-of-zelda/50876050. Added functionality so there is a delay in enemies taking damage for the collision. This makes it so enemies dont die on one hit every time, since collisions are updated and registered per frame.