Begin work on enemy item drops
Enemy drops will be based on the classification of an enemy, and on a counter that is updated every time an enemy is defeated, beginning at 0 and going up to 9. Refer to the chart at the  links below for the schema. I will not be implementing (unless we decide we want to) a percent chance of an item being dropped in the first place, just to make the game a little easier to play. Note that in the actual game, an item does not always drop, since each classification of enemy has a certain percentage chance to drop an item. Also, I will be including code for class A enemies, just for the sake of expandability in the codebase, not necessarily that we have any class A enemies. I will however be excluding functionality for class x, since those enemies do not ever drop items. I will include it for the sake of classification, since again, if we were to expand this game further than the class requirements, it would be good to know which enemies are not supposed to drop items. https://www.zeldaspeedruns.com/loz/generalknowledge/item-drops-chart. https://www.zeldadungeon.net/zelda-runners-examining-random-and-forced-drops-and-chatting-with-zant/